feat: Frontend logic for CRUD operations

+++ controllers/goalController.js
@@ -4,26 +4,42 @@ const User = require("../models/User");
 const createGoal = async (req, res) => {
     try {
         const { goalName, description } = req.body;
+        const username = req.user.username;  // Extract username from decoded JWT
 
-        const newGoal = await new Goal({ goalName, description, userId: req.user._id });
+        const user = await User.findOne({ username });
+        if (!user) {
+            return res.status(404).json({ error: "User not found" });
+        }
+
+        const userId = user._id;  // Extract id from user object
+
+        if (!goalName || !description) {
+            return res.status(400).json({ error: "Missing required fields" });
+        }
+
+        const newGoal = new Goal({ goalName, description, userId });
         await newGoal.save();
 
         // Add goal reference to user's goals array
-        await User.findByIdAndUpdate(req.user._id,
-                                    { $push: { goals: newGoal._id }},
-                                    { new: true }
+        await User.findByIdAndUpdate(userId,
+            { $push: { goals: newGoal._id }},
+            { new: true }
         );
 
         res.status(201).json(newGoal);
     } catch (error) {
         console.error(error);
-        res.status(500).json({ error: "Failed to craete a new goal" });
+        res.status(500).json({ error: "Failed to create a new goal" });
     }
 }
 
 const readGoals = async (req, res) => {
     try {
-        const goals = await Goal.find({ userId: req.user._id }).sort({ createdAt: -1 });
+        const username = req.user.username;
+        const user = await User.findOne({ username });
+        const userId = user._id;
+
+        const goals = await Goal.find({ userId }).sort({ createdAt: -1 });
         res.status(200).json(goals);
     } catch (error) {
         console.error(error);
@@ -33,12 +49,20 @@ const readGoals = async (req, res) => {
 
 const updateGoal = async (req, res) => {
     try {
-        const { name } = req.params;
-        const { updateGoalName, descriptionUpdate } = req.body;
+        const username = req.user.username;
+        const user = await User.findOne({ username });
+        const userId = user._id;
+
+        const { updateGoalByName } = req.body;
+        const { updatedGoalName, descriptionUpdate } = req.body;
+
+        if (!updatedGoalName || !descriptionUpdate) {
+            return res.status(400).json({ error: "Missing required fields" });
+        }
 
         const updatedGoal = await Goal.findOneAndUpdate(
-            { goalName: name, userId: req.user._id },
-            { goalName: updateGoalName, description: descriptionUpdate },
+            { goalName: updateGoalByName, userId },
+            { goalName: updatedGoalName, description: descriptionUpdate },
             { new: true }
         );
 
@@ -55,11 +79,21 @@ const updateGoal = async (req, res) => {
 
 const deleteGoal = async (req, res) => {
     try {
-        const { name } = req.params;
+        const username = req.user.username;
+        const user = await User.findOne({ username });
+        const userId = user._id;
 
-        const deletedGoal = await Goal.findOneAndDelete({ 
-            goalName: name,
-            userId: req.user._id
+        const { deleteGoalByName } = req.body;
+
+        const deletedGoal = await Goal.findOneAndDelete({ goalName: deleteGoalByName, userId });
+
+        if (!deletedGoal) {
+            return res.status(404).json({ error: "Goal not found" });
+        }
+
+        // Delete goal reference to user's goals array
+        await User.findByIdAndDelete(userId, {
+            $pull: { goals: deletedGoal._id }
         });
 
         res.status(200).json(deletedGoal);
